upload

+++ app/controllers/application-list.js
@@ -5,8 +5,10 @@ export default Controller.extend({
 
     actions: {
       destroyApplication(destroyId) {
-        let application = this.store.peekRecord("application", destroyId)
-        application.destroyRecord()
+        if (confirm("Are you sure want to delete this record ? ") == true) {
+          let application = this.store.peekRecord("application", destroyId)
+          application.destroyRecord()
+        } 
       }
     }
   });
